Trait embre_crypt::aes::AESStrCompare

pub trait AESStrCompare {
    // Required methods
    fn aes_compare_slice(
        &self,
        encrypted: &[u8],
        key: &[u8],
        iv: Option<&[u8]>,
        plaintext: &[u8]
    ) -> bool;
    fn aes_compare_w_str(
        &self,
        encrypted: &[u8],
        key: &[u8],
        iv: Option<&[u8]>,
        plaintext: &[u16]
    ) -> bool;
}

Separate trait for string compare functions. The implementation for DefaultAESCrypter checks if the result would be the same length, first. Then it uses a local buffer, chunks the string into slices the same size as the block size, and encrypts the slices into the local buffer and compares the part of the encrypted chunk to the newly encrypted chunk of data, until it gets through the whole string. It does the same for the wide string implementation, except it uses an additional buffer to write out the UTF-8 version of the wide string character to an additional buffer for comparison.

Required Methods

fn aes_compare_slice( &self, encrypted: &[u8], key: &[u8], iv: Option<&[u8]>, plaintext: &[u8] ) -> bool

Compares a string with the encrypted data. The implementation of this on the DefaultAESCrypter assumes the encrypted and plaintext string is UTF-8.

Arguments

encrypted: &[u8] - Encrypted bytes.

key: &[u8] - AES Key for encrypted bytes.

iv: Option<&[u8]> - Optional IV for encrypted bytes.

plaintext: &[u8] - The plaintext UTF-8 string we are comparing against.

returns: bool

fn aes_compare_w_str( &self, encrypted: &[u8], key: &[u8], iv: Option<&[u8]>, plaintext: &[u16] ) -> bool

Compares a wide string with the encrypted data. The implementation of this on the DefaultAESCrypter assumes the encrypted string is UTF-8, so it writes the ASCII version of each character of the ‘plaintext’ wide string to an additional buffer for the encryption and comparison.

Arguments

encrypted: &[u8] - Encrypted bytes.

key: &[u8] - AES Key for encrypted bytes.

iv: Option<&[u8]> - Optional IV for encrypted bytes.

plaintext: &[u16] - The plaintext wide string we are comparing against.

returns: bool